+++ fabric/src/main/java/io/github/sakurawald/core/structure/descriptor/MetaDescriptor.java
@@ -31,4 +31,9 @@ public class MetaDescriptor<T> extends StringDescriptor{
     public int sortPriority() {
         return +1;
     }
+
+    @Override
+    public String getStringType() {
+        return "Meta";
+    }
 }