+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
 - [core] feature: new `print_user_guide` debug option in `fuji/config.json` file.
+- [fuji] feature: now will display the `string type` for `/fuji inspect permissions-and-metas` command.
+